split tests from common code
